Added give 100 code on left mouse click
This adds the functionality to give the player 100 gold when they click the left mouse button in game. This accomplishes this by mapping the game structures for the game UI, game character, and game client, and the respective methods that perform that are needed to perform this action. This serves as the first functionality of the cheat.