Add check for pause menu befroe giving 100 gold.
This adds the apporpriate member function to the CGameUI class to check if the game is paused. This was added to ensure that we do not receive gold whenever clicking back into the game. Later down the road I can simply add a seperate menu that toggles the gold amount to enter.